refactor(cli): move data commands into megatron backend
